Revert "Merge branch 'dev-model' into main"
This reverts commit f5196510cb45c10212e6061752a3dfaad5cfc24f, reversing
changes made to 1a796f4a4d1c056b3febe1259a244188e9d68372.
