+++ backend/requirements.txt
@@ -1,10 +1,3 @@
-pytest==7.2.0
 Flask==2.2.2
 Flask-Cors==3.0.10
 requests==2.28.1
-numpy==1.23.4
-pandas==1.5.1
-scikit-learn==1.1.3
-scipy==1.9.3
-spotipy==2.21.0
-tqdm==4.64.1
