--- backend/src/model/.gitignore
@@ -1,5 +1,0 @@
-.git/
-.idea/
-__pycache_/
-venv/
-.cache
