+++ frontend/tests/components/SpotifySearch.test.tsx
@@ -1,11 +1,11 @@
-import { QueryClientProvider } from '@tanstack/react-query'
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { render } from '@testing-library/react'
 
-import { queryClient } from 'lib/router'
 import SpotifySearch from 'components/SpotifySearch'
 
 describe('test SpotifySearch.tsx component', () => {
   test('renders input field', () => {
+    const queryClient = new QueryClient()
     const { getByRole } = render(
       <QueryClientProvider client={queryClient}>
         <SpotifySearch />
