+++ frontend/tsconfig.json
@@ -21,9 +21,8 @@
     "baseUrl": ".",
     "paths": {
       "src/*": ["src/*"],
-      "components/*": ["src/components/*"],
       "lib/*": ["src/lib/*"],
-      "pages/*": ["src/pages/*"]
+      "components/*": ["src/components/*"]
     }
   },
   "include": ["vite.config.ts", "**/*.ts", "**/*.tsx"],
